requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-chroma==0.1.4
langchain-core==0.3.15
langchain-text-splitters==0.3.2
chromadb==0.5.11
python-dotenv==1.0.1
pypdf==4.3.1